# drustvena_mreza/Web App/api/api.yaml
swagger: '2.0'
info:
  title: Bubbles API
  description: Mrzim svoj život.
  version: 1.0.0
produces:
  - application/json
paths:
...
  /user/isContact:
    get:
      tags:
        - User
      summary: Provjerava jel osoba sa id-em id prijatelj vlasnika session-a
      parameters:
        - name: id
          in: query
          required: true
          type: integer
      responses:
        '200':
          description: Nakon provjere
          schema:
            type: object
            properties:
              isContact:
                type: boolean
                description: true ako je osoba sa id-em id prijatelj vlasnika session-a
  /user/info:
    get:
      tags:
        - User
      summary: Dohvaća podatke o korisniku s id-em id
      parameters:
        - name: id
          in: query
          required: true
          type: integer
      responses:
        '200':
          description: Uspješan dohvat
          schema:
            $ref: '#/definitions/User'
  /user/contactRequest:
    post:
      tags:
        - User
      summary: Vlasnik session-a dodaje osobu sa id-em id za prijatelja (može gledati njegove stvari)
      parameters:
        - name: user_id
          in: formData
          required: true
          type: integer
      responses:
        '200':
          description: Uspješno dodavanje
  /user/contacts:
    get:
      tags:
        - User
      summary: Vraća kontakte od korisnika s id-em id ili od vlasnika sessiona ako id nije pružen
      parameters:
        - name: id
          in: query
          required: false
          type: integer
      responses:
        '200':
          description: Uspješno dohvaćanje
          schema:
            type: object
            properties:
              contacts:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    username:
                      type: string
...
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        format: int32
      username:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      country_id:
        type: string
      city:
        type: string
      address:
        type: string
      relationship_status_id:
        type: integer
      gender_id:
        type: integer
...